train/orpo_train.py
"""ORPO trainer for Clinical Note Summarizer.

Fine-tunes BioMistral-7B with QLoRA (r=32, alpha=64) and 4-bit NF4 quantization
using ORPO preference optimization via TRL ORPOTrainer.

Usage:
    python train/orpo_train.py --config configs/orpo_base.yaml
"""

import argparse
import os

import yaml
from dotenv import load_dotenv
from datasets import load_dataset
from transformers import (
    AutoModelForCausalLM,
    AutoTokenizer,
    BitsAndBytesConfig,
)
from peft import LoraConfig, get_peft_model, prepare_model_for_kbit_training
from trl import ORPOConfig, ORPOTrainer
import wandb

load_dotenv()


def load_config(config_path: str) -> dict:
    """Load YAML config and resolve environment variable references."""
    raise NotImplementedError


def build_quantization_config() -> BitsAndBytesConfig:
    """Build 4-bit NF4 quantization config for BioMistral."""
    raise NotImplementedError


def build_lora_config(cfg: dict) -> LoraConfig:
    """Build QLoRA config (r=32, alpha=64) from YAML settings."""
    raise NotImplementedError


def load_model_and_tokenizer(cfg: dict):
    """Load BioMistral-7B with quantization and prepare for k-bit training."""
    raise NotImplementedError


def load_data(cfg: dict):
    """Load ORPO preference pairs from JSONL splits."""
    raise NotImplementedError


def train(cfg: dict) -> None:
    """Orchestrate ORPO training: model loading, data prep, ORPOTrainer run."""
    raise NotImplementedError


def main() -> None:
    """Entry point: parse --config argument and launch training."""
    parser = argparse.ArgumentParser(description="ORPO trainer for Clinical Note Summarizer")
    parser.add_argument("--config", type=str, required=True, help="Path to YAML config file")
    parser.add_argument("--resume", type=str, default=None, help="Path to checkpoint to resume from")
    args = parser.parse_args()

    cfg = load_config(args.config)
    if args.resume:
        cfg["resume_from_checkpoint"] = args.resume
    train(cfg)


if __name__ == "__main__":
    main()
