requirements.txt
transformers==4.44.0
peft==0.12.0
trl==0.9.6
bitsandbytes==0.43.3
datasets>=2.20.0
accelerate>=0.33.0
wandb>=0.17.0
fastapi>=0.111.0
uvicorn>=0.30.0
streamlit>=1.36.0
bert-score>=0.3.13
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.2.0
